agente_videos_multi_ia/agente_videos/requirements.txt
openai>=1.12.0
requests>=2.31.0
pyyaml>=6.0.1
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pandas>=2.2.0
loguru>=0.7.2
tenacity>=0.8.2
python-dotenv>=1.0.1
jinja2>=3.1.3
